largo-lagrande-lambda/functions/classify/requirements.txt
boto3==1.28.0
botocore==1.31.0
pyyaml==5.4.1
